Melhorado Exercícios do dia 06/04 Finalizado dia 16/04/2022. -------------- COISAS A SEREM IMPLEMENTADAS/MELHORADAS ------------- EX6: Não consegui ordenar por naipe e por valor EX7: Não há tratamento dos produtos e itens caso eles não coexistirem, o mesmo com a quantidade EX1: O banco de dados da agenda está sendo tratado estaticamente, sem possibilidade de exclusão da memória.
--------------------- FORAM ADICIONADAS ----------------------
EX6: ORDENAÇÃO PELO VALOR DA CARTA
EX6: GERAÇÃO AUTOMÁTICA DA CARTA
EX8: CELULAR COM BATERIA ATÉ 0
EX7: PREÇO EM DOUBLE E NÃO FLOAT

+++ 06.04/src/Ex7/Caixa.java
@@ -22,8 +22,8 @@ public class Caixa {
         }
         System.out.println("-----------------------------------------------------");
     }
-    public float buscandoPrecoItem(Produto[] produtos, Item[] itens, int qtdPedido, int qtdProdutos ) {
-        float preco = 0.0f;
+    public double buscandoPrecoItem(Produto[] produtos, Item[] itens, int qtdPedido, int qtdProdutos ) {
+        double preco = 0.0f;
 
         for (int i=0; i < qtdPedido; i++) {
             for (int j=0; j < qtdProdutos; j++) {
@@ -36,10 +36,10 @@ public class Caixa {
         System.out.println("Não foi encontrado um produto");
         return preco;
     }
-    public float valorTotalCompra(Produto[] produtos, Item[] itens, int qtdPedido, int qtdProdutos) {
-        float total = 0.0f;
+    public double valorTotalCompra(Produto[] produtos, Item[] itens, int qtdPedido, int qtdProdutos) {
+        double total = 0.0f;
 
-        float precoItem = buscandoPrecoItem(produtos, itens, qtdPedido, qtdProdutos);
+        double precoItem = buscandoPrecoItem(produtos, itens, qtdPedido, qtdProdutos);
         if (precoItem > 0.0) {
             total += precoItem * qtdPedido;
         }
@@ -47,7 +47,7 @@ public class Caixa {
     }
     public void pagandoPedidos(Produto[] produtos, Item[] itens, int qtdPedido, int qtdProdutos, Scanner keyboard) {
         String wordPagamento = ""; int opPagamento;
-        float valorTotal =  valorTotalCompra(produtos, itens, qtdPedido, qtdProdutos);
+        double valorTotal =  valorTotalCompra(produtos, itens, qtdPedido, qtdProdutos);
         System.out.println("O TOTAL DA COMPRA FOI DE " + valorTotal + " REAIS");
         System.out.println("[1] PAGAR NO CARTAO CREDITO\t[2] PAGAR NO CARTAO DEBITO\t[3] PAGAR NO DINHEIRO\t[4] PAGAR NO PIX");
         opPagamento = keyboard.nextInt();
